inference/pyproject.toml
[project]
name = "inference-service"
version = "0.3.0"
description = "Inference service for SagePick with agentic movie recommendations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.11.0",
    "wandb>=0.18.0",
    "joblib>=1.4.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "google-generativeai>=0.8.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "tavily-python>=0.5.0",
]
